+++ src/main/java/Duke/EventHandler.java
@@ -1,3 +1,5 @@
+package Duke;
+
 import java.util.ArrayList;
 
 public class EventHandler {
