+++ src/main/java/Duke/TaskList.java
@@ -1,3 +1,5 @@
+package Duke;
+
 public class TaskList {
 
 }
