+++ src/main/java/Duke/UserInterface.java
@@ -1,3 +1,5 @@
+package Duke;
+
 import java.util.ArrayList;
 
 public class UserInterface {
