--- src/main/java/Event.java
@@ -1,13 +1,0 @@
-public class Event extends Task{
-    protected String at;
-
-    public Event(String description, String at) {
-        super(description);
-        this.at = at;
-    }
-
-    @Override
-    public String toString() {
-        return "[E]" + super.toString() + "(at: " + at + ")";
-    }
-}
